+++ server/src/auth/auth_oidc.js
@@ -30,7 +30,7 @@ exports.initialize = async () =>{
       logger.info("OIDC is not enabled")
     }else{
       oidcConfig = await OIDC.find()
-      settings = await Settings.find()
+      settings = await Settings.findUrl()
       url = settings.url?.replace(/\/$/g,'')
       if(!url){
         logger.error("AnsibleForms Url is not set")