+++ server/src/db/create_schema_and_tables.sql
@@ -114,7 +114,8 @@ CREATE TABLE `settings` (
   `mail_username` varchar(250) DEFAULT NULL,
   `mail_password` text DEFAULT NULL,
   `mail_from` varchar(250) DEFAULT NULL,
-  `url` varchar(250) DEFAULT NULL
+  `url` varchar(250) DEFAULT NULL,
+  `forms_yaml` longtext CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci DEFAULT NULL 
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 USE `AnsibleForms`;
 DROP TABLE IF EXISTS `azuread`;