+++ server/src/lib/mssql.js
@@ -2,53 +2,58 @@
 const logger = require('./logger');
 const client = require('mssql');
 const Credential = require('../models/credential.model')
+const Helpers = require('./common')
 
 Mssql = {}
 
-Mssql.query=async function(connection_name,query){
-  return Credential.findByName(connection_name)
-  .then((creds)=>{
-    var config = {
-        server: creds.host,
-        user: creds.user,
-        password: creds.password,
-        database: creds.db_name,
-        port: creds.port,
-        options: {
-            trustServerCertificate: true
-        }
-    };
-    if(creds.secure){
-      
-      config.encrypt=true
-    }
-    // remove database if needed
-    if(!creds.db_name){
-      delete config.database
-    }
-    return config
-  })
-  .then((config)=>{
-    return new Promise((resolve,reject)=>{
-      client.connect(config, function(err,conn){
-        if(err){
-          reject(`[${connection_name}] connection error : ${err.toString()}`)
-        }else{
-          conn.query(query, (err, result) => {
-            // close connection immediately
-            logger.debug(`[${connection_name}] closing connection`)
-            conn.close()
-            if(err){
-              reject(`[${connection_name}] query error : ${err.toString()}`)
-            }
-            resolve(result?.recordset)
-          })
-        }
+// rewrite with await 5.0.3
 
-      })
-    })
+Mssql.query = async function (connection_name, query) {
 
-  })
-};
+  var creds = await Credential.findByName(connection_name)
+  var config = {
+    server: creds.host,
+    user: creds.user,
+    password: creds.password,
+    database: creds.db_name,
+    port: creds.port,
+    options: {
+      trustServerCertificate: true
+    }
+  };
+  if(creds.secure){
+    
+    config.encrypt=true
+  }
+  // remove database if needed
+  if(!creds.db_name){
+    delete config.database
+  }
+
+  var conn
+  try{
+    var conn = await client.connect(config)
+  }catch(err){
+    logger.error(`[${connection_name}] connection error`,err)
+    throw err
+  }
+
+  var result
+  try{
+    result = await conn.query(query)
+    return result?.recordset
+  }catch(err){
+    logger.error(`[${connection_name}] query error`,err)
+    throw err
+  }finally{
+    try{
+      conn.close()
+    }catch(e){
+      logger.error(`[${connection_name}] connection error`,e)
+      throw e
+    }
+  }
+
+}
 
 module.exports = Mssql